brightbridgeDir/requirements.txt
google-cloud-aiplatform[adk,agent_engines]>=1.49.0
google-adk>=1.0.0
python-dotenv==1.1.0
litellm
streamlit>=1.31.0
requests>=2.31.0
fastapi>=0.100.0
uvicorn[standard]>=0.20.0